Titanic Notebook updated
The Linear Model training and testing was (is) so easy with pandas and sklearn... When you try to reinvent the wheel with pyspark and Zeppelin itself, a bunch of things go wrong everywhere. It's just a matter of time to get this right, though.

I got the feeling I'll have to invest my whole weekend in this thing. Oh boy.