Linear Regression completed; visualizations added
Told you I got this! There's a catch, though: the code for the pyspark's version of the Linear Regression works for a newer version of pyspark than the one that comes bundled with Apache Zeppelin. And guess what: you can't change it without breaking Zeppelin's %spark.pyspark interpreter! It works without any troubles on a Jupyter notebook. Some things never change...

In addition, I inserted a bunch of visualizations (prior and after age imputing through the Linear Regression model) aiming to define the most dominant characteristics of the deceased and survivor groups (class, age and sex, so far). I'll try to concatenate again the Survived column and remake the Age histogram with the imputed data later.